+++ packages/mobile-visualization/CHANGELOG.md
@@ -8,6 +8,12 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## 3.4.0-beta.24 (3/12/2026 PST)
+
+#### 🚀 Updates
+
+- Improve chart accessibility. [[#492](https://github.com/coinbase/cds/pull/492)]
+
 ## 3.4.0-beta.23 (3/10/2026 PST)
 
 #### 🚀 Updates