+++ packages/ui-mobile-playground/CHANGELOG.md
@@ -8,6 +8,12 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## 4.12.0 (3/11/2026 PST)
+
+#### 🚀 Updates
+
+- Update mobile routes. [[#492](https://github.com/coinbase/cds/pull/492)]
+
 ## 4.11.0 (3/10/2026 PST)
 
 #### 🚀 Updates