+++ packages/web-visualization/CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## 3.4.0-beta.24 (3/12/2026 PST)
+
+This is an artificial version bump with no new change.
+
 ## 3.4.0-beta.23 (3/10/2026 PST)
 
 #### 🚀 Updates